+++ front/src/routes/(app)/admin/+page.ts
@@ -1,0 +1,11 @@
+// Copyright (c) 2025 Start Codex SAS. All rights reserved.
+// SPDX-License-Identifier: BUSL-1.1
+
+import { redirect } from '@sveltejs/kit';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ parent }) => {
+	const { user } = await parent();
+	if (!user?.is_instance_admin) redirect(302, '/');
+	redirect(302, '/admin/smtp');
+};